requirements.txt
streamlit
st-supabase-connection
supabase
streamlit_calendar
python-dotenv